+++ google-cloud-spanner/src/main/java/com/google/cloud/spanner/GrpcResultSet.java
@@ -90,6 +90,8 @@ class GrpcResultSet extends AbstractResultSet<List<Object>> implements ProtobufR
       boolean hasNext = currRow.consumeRow(iterator);
       if (!hasNext) {
         statistics = iterator.getStats();
+        // Close the ResultSet when there is no more data.
+        close();
       }
       return hasNext;
     } catch (Throwable t) {
@@ -113,9 +115,14 @@ class GrpcResultSet extends AbstractResultSet<List<Object>> implements ProtobufR
 
   @Override
   public void close() {
+    synchronized (this) {
+      if (closed) {
+        return;
+      }
+      closed = true;
+    }
     listener.onDone(iterator.isWithBeginTransaction());
     iterator.close("ResultSet closed");
-    closed = true;
   }
 
   @Override