+++ google-cloud-spanner/src/main/java/com/google/cloud/spanner/SessionImpl.java
@@ -16,6 +16,7 @@
 
 package com.google.cloud.spanner;
 
+import static com.google.cloud.spanner.SessionClient.optionMap;
 import static com.google.cloud.spanner.SpannerExceptionFactory.newSpannerException;
 
 import com.google.api.core.ApiFuture;
@@ -27,6 +28,7 @@ import com.google.cloud.spanner.AbstractReadContext.SingleReadContext;
 import com.google.cloud.spanner.AbstractReadContext.SingleUseReadOnlyTransaction;
 import com.google.cloud.spanner.Options.TransactionOption;
 import com.google.cloud.spanner.Options.UpdateOption;
+import com.google.cloud.spanner.SessionClient.SessionOption;
 import com.google.cloud.spanner.TransactionRunnerImpl.TransactionContextImpl;
 import com.google.cloud.spanner.spi.v1.SpannerRpc;
 import com.google.common.base.Ticker;
@@ -92,19 +94,38 @@ class SessionImpl implements Session {
 
     /** Registers the current span on the transaction. */
     void setSpan(ISpan span);
+
+    /** Closes the transaction. */
+    void close();
   }
+
+  static final int NO_CHANNEL_HINT = -1;
 
   private final SpannerImpl spanner;
   private final SessionReference sessionReference;
   private SessionTransaction activeTransaction;
   private ISpan currentSpan;
   private final Clock clock;
+  private final Map<SpannerRpc.Option, ?> options;
 
   SessionImpl(SpannerImpl spanner, SessionReference sessionReference) {
+    this(spanner, sessionReference, NO_CHANNEL_HINT);
+  }
+
+  SessionImpl(SpannerImpl spanner, SessionReference sessionReference, int channelHint) {
     this.spanner = spanner;
     this.tracer = spanner.getTracer();
     this.sessionReference = sessionReference;
     this.clock = spanner.getOptions().getSessionPoolOptions().getPoolMaintainerClock();
+    this.options = createOptions(sessionReference, channelHint);
+  }
+
+  static Map<SpannerRpc.Option, ?> createOptions(
+      SessionReference sessionReference, int channelHint) {
+    if (channelHint == NO_CHANNEL_HINT) {
+      return sessionReference.getOptions();
+    }
+    return optionMap(SessionOption.channelHint(channelHint));
   }
 
   @Override
@@ -113,7 +134,7 @@ class SessionImpl implements Session {
   }
 
   Map<SpannerRpc.Option, ?> getOptions() {
-    return sessionReference.getOptions();
+    return options;
   }
 
   void setCurrentSpan(ISpan span) {
@@ -440,6 +461,10 @@ class SessionImpl implements Session {
         .build();
   }
 
+  SessionTransaction getActiveTransaction() {
+    return this.activeTransaction;
+  }
+
   <T extends SessionTransaction> T setActive(@Nullable T ctx) {
     throwIfTransactionsPending();
     // multiplexed sessions support running concurrent transactions
@@ -454,6 +479,12 @@ class SessionImpl implements Session {
     }
     return ctx;
   }
+
+  void onError(SpannerException spannerException) {}
+
+  void onReadDone() {}
+
+  void onTransactionDone() {}
 
   TraceWrapper getTracer() {
     return tracer;